% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gammak.max.R
\name{get_gammak.max}
\alias{get_gammak.max}
\title{Computes an upper bound for the group lasso regularization penalty.}
\usage{
get_gammak.max(data, group.length, k, type, lambda, beta = rep(0,
  ncol(data$x)), beta0 = 0)
}
\arguments{
\item{data}{A list with x and y}

\item{group.length}{A vector of group lengths}

\item{k}{A value between 1 and the number of groups}

\item{type}{String: \code{'linear'} or \code{'logit'}}

\item{lambda}{A vector of length 2, with the values of \eqn{\lambda_1} and \eqn{\lambda_2}}

\item{beta}{A vector, with an approximate \eqn{\beta}}

\item{beta0}{The intercept}
}
\value{
A number. The upper bound for \eqn{\gamma_k}.
}
\description{
Computes an upper bound of \eqn{\gamma_k} so that the solution is non zero.
}
