% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lambda1.max.R
\name{get_lambda1.max}
\alias{get_lambda1.max}
\title{Computes the maximum weight of the lasso penalty}
\usage{
get_lambda1.max(data, group.length, beta = rep(0, ncol(data$x)), beta0 = 0,
  type)
}
\arguments{
\item{data}{A list with x and y}

\item{group.length}{A vector with group lengths}

\item{beta}{A vector, the previous solution \eqn{\beta}}

\item{beta0}{A number, the intercept}

\item{type}{A string. Type of loss function: \code{'logit'} or \code{'linear'}}
}
\value{
A number. The maximum \eqn{\lambda_1}
}
\description{
Given the \code{data} and \eqn{\beta}, this function computes an upper bound for \eqn{\lambda_1} such that \eqn{\beta(\lambda_1)\ne 0}.
}
