% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_validation_cost.R
\name{get_validation_cost}
\alias{get_validation_cost}
\title{Evaluates the risk function}
\usage{
get_validation_cost(X, y, model_params, type = "linear")
}
\arguments{
\item{X}{The model matrix}

\item{y}{The response vector}

\item{model_params}{The sgl fit.}

\item{type}{String: \code{'linear'} or \code{'logit'}.}
}
\value{
The sum of errors
}
\description{
Evaluates the risk function
}
