% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.length2index.R
\name{group.length2index}
\alias{group.length2index}
\title{Workaround to convert group lengths to index vector for the sparse-group lasso.}
\usage{
group.length2index(group.length)
}
\arguments{
\item{group.lenght}{Vector with group lengths}
}
\value{
Index vector
}
\description{
\code{group.length2index} converts group lengths, e.g. \code{[1,2,3]} to \code{[1,2,2,3,3,3]}.
}
\seealso{
index2group.length
}
