% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index2group.length.R
\name{index2group.length}
\alias{index2group.length}
\title{Workaround to convert index vector to group lengths.}
\usage{
index2group.length(index)
}
\description{
Workaround to convert index vector to group lengths.
}
\seealso{
group.length2index
}
