% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isgl.R
\name{isgl}
\alias{isgl}
\title{The iterative sparse-group lasso}
\usage{
isgl(data.train, data.validate, index = NULL, group.length = NULL,
  type = "linear", standardize = F, momentum = 2)
}
\arguments{
\item{data.train}{A list with the matrix of observations \code{x}, and the response \code{y}, to train the model.}

\item{data.validate}{A list with the matrix of observations \code{x}, and the response \code{y}, to validate the model.}

\item{index}{A vector with group indices.}

\item{group.length}{A vector with group length, alternatively to \code{index}.}

\item{type}{String: \code{'linear'} or \code{'logit'}.}

\item{standardize}{Whether to stardardize \code{data.train} and \code{data.validate}. Recommended if \code{type} is \code{'logit'}.}

\item{momentum}{Acceleration rate. Should be > 1.}
}
\value{
An object of class \code{isgl}.
}
\description{
Computes the solution of the sparse-group lasso problem, defined as,
\deqn{\hat{\beta}(\lambda, \gamma)=\arg\min_{\beta\in B} \left\{ \hat{R}(\beta) + \lambda_2 \sum_{j=1}^{J} \gamma_j \|\beta^{(j)}\|_2 +  \lambda_1 \|\beta\|_1 \right\}.}
Regularization parameters \eqn{\lambda} and \eqn{\gamma} are automatically selected, so that the validation error is minimized.
}
