% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.isgl.R
\name{predict.isgl}
\alias{predict.isgl}
\title{Predicts new response, for an object of class \code{isgl}.}
\usage{
\method{predict}{isgl}(isgl, newX)
}
\arguments{
\item{isgl}{Object of class \code{isgl}.}

\item{newX}{New observations, as supplied to the isgl algorithm.}
}
\value{
Vector of predictions.
}
\description{
Predicts new response, for an object of class \code{isgl}.
}
\seealso{
isgl
}
