% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGL.R
\name{SGLinner}
\alias{SGLinner}
\title{Function to compute the SGL solution}
\usage{
SGLinner(data, group.length, type = "linear", maxit = 1000,
  thresh = 0.001, min.frac = 0.1, gamma = 0.8, verbose = FALSE,
  step = 1, reset = 10, lambda1 = 0.2, lambda2 = 0.01,
  groupW = NULL, betas.old = rep(0, ncol(data$x)), weights = rep(1,
  NROW(data$x)))
}
\description{
Originally written by Simon for R package SGL.
}
