% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_inner_problem.R
\name{solve_inner_problem}
\alias{solve_inner_problem}
\title{Solves the sparse-group lasso problem}
\usage{
solve_inner_problem(data.train, group.length, lambdas, type = "linear",
  betas.old = rep(0, ncol(data.train$x)), simple = F)
}
\arguments{
\item{data.train}{A list with matrix \code{x} and response vector \code{y}.}

\item{group.length}{A vector of group lengths. Columns of \code{x} are in the same order of groups.}

\item{lambdas}{A vector of length \eqn{J+2}. The first two components are those of \eqn{\lambda}, and the remaining components are the group weights \eqn{\gamma}.}

\item{type}{A string. Whether \code{'linear'} or \code{'logit'}}

\item{betas.old}{A vector, with an approximate initial solution \eqn{\beta}. Be careful!}

\item{simple}{Boolean, indicating whether we are solving the two-parameter version.}
}
\value{
A list with the fit \code{beta} and \code{intercept}.
}
\description{
Computes the solution of the SGL problem, defined as,
\deqn{\hat{\beta}(\lambda, \gamma)=\arg\min_{\beta\in B} \left\{ \hat{R}(\beta) + \lambda_2 \sum_{j=1}^{J} \gamma_j \|\beta^{(j)}\|_2 +  \lambda_1 \|\beta\|_1 \right\}.}
}
