% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standarize_initial_data.R
\name{standarize_inital_data}
\alias{standarize_inital_data}
\title{Standarize the data for the sparse-group lasso}
\usage{
standarize_inital_data(data.train, data.validate, type = "linear")
}
\arguments{
\item{data.train}{A list with matrix x and response y}

\item{data.validate}{A list with matrix x and response y}
}
\value{
A list with the transformed data, the transformation,
and intercept.
}
\description{
Standarize the data for the sparse-group lasso
}
